Use light DOM to render static pages.
