+++ client/elements/sc-site-layout.js
@@ -1,4 +1,4 @@
-import { LitElement, html, css } from 'lit';
+import { LitElement, html } from 'lit';
 
 import { BrowserMicroSentryClient } from '@micro-sentry/browser';
 
@@ -8,6 +8,7 @@ import { icon } from '../img/sc-icon';
 import './sc-page-selector';
 import './navigation/sc-navigation-linden-leaves';
 import './addons/sc-linear-progress';
+import './menus/sc-menu-static-pages-nav';
 
 import { LitLocalized } from './addons/sc-localization-mixin';
 import { store } from '../redux-store';
@@ -22,8 +23,8 @@ const microSentryClient = new BrowserMicroSentryClient({
 });
 
 class SCSiteLayout extends LitLocalized(LitElement) {
-  static get styles() {
-    return css``;
+  createRenderRoot() {
+    return this;
   }
 
   render() {
@@ -31,6 +32,7 @@ class SCSiteLayout extends LitLocalized(LitElement) {
       <style>
         ${SCSiteLayoutStyles}
       </style>
+
       <div id="universal_toolbar">
         <sc-navigation-linden-leaves id="breadCrumb"></sc-navigation-linden-leaves>
 
@@ -43,179 +45,13 @@ class SCSiteLayout extends LitLocalized(LitElement) {
             <div id="subTitle">${this.localize('interface:pageSubtitle')}</div>
           </div>
         </div>
-        <sc-linear-progress .active="${this.linearProgressActive}"></sc-linear-progress>
 
-        <div id="static_pages_nav_menu">
-          <nav>
-            <ul>
-              ${this.toolbarSelectedTemplate} ${this.shouldShowSecondToolbarTemplate}
-              ${this.shouldShowTipitakaToolbarTemplate} ${this.shouldShowAcademicToolbarTemplate}
-              ${this.shouldShowOrganizationalToolbarTemplate}
-              ${this.shouldShowGuidesToolbarTemplate}
-            </ul>
-          </nav>
-        </div>
+        <sc-linear-progress .active=${this.linearProgressActive}></sc-linear-progress>
+        <sc-menu-static-pages-nav id="static_pages_nav_menu"></sc-menu-static-pages-nav>
       </div>
 
       <sc-page-selector id="page_selector"></sc-page-selector>
     `;
-  }
-
-  createRenderRoot() {
-    return this;
-  }
-
-  get toolbarSelectedTemplate() {
-    return html`
-      ${this.staticPagesToolbarDisplayState &&
-      this.staticPagesToolbarDisplayState.displayFirstToolbar
-        ? html`
-            <li>
-              <a href="/introduction">${this.localize('interface:introduction')}</a>
-            </li>
-            <li>
-              <a href="/donations">${this.localize('interface:donations')}</a>
-            </li>
-            <li>
-              <a href="/offline">${this.localize('interface:useOffline')}</a>
-            </li>
-            <li>
-              <a
-                href="https://discourse.suttacentral.net/c/meta/updates"
-                class="external"
-                title="See updates on SuttaCentral forum"
-                target="_blank"
-                rel="noopener"
-              >
-                ${this.localize('interface:whatsnew')}
-              </a>
-            </li>
-          `
-        : ''}
-    `;
-  }
-
-  get shouldShowSecondToolbarTemplate() {
-    return html`
-      ${this.staticPagesToolbarDisplayState &&
-      this.staticPagesToolbarDisplayState.displaySecondToolbar
-        ? html`
-            <li>
-              <a href="/subjects">${this.localize('interface:subjects')}</a>
-            </li>
-            <li>
-              <a href="/similes">${this.localize('interface:similes')}</a>
-            </li>
-            <li>
-              <a href="/names">${this.localize('interface:names')}</a>
-            </li>
-            <li>
-              <a href="/terminology">${this.localize('interface:terminology')}</a>
-            </li>
-          `
-        : ''}
-    `;
-  }
-
-  get shouldShowTipitakaToolbarTemplate() {
-    return html`
-      ${this.staticPagesToolbarDisplayState &&
-      this.staticPagesToolbarDisplayState.displayTipitakaToolbar
-        ? html`
-            <li>
-              <a href="/discourses-guide-sujato">${this.localize('interface:discourses')}</a>
-            </li>
-            <li>
-              <a href="/vinaya-guide-brahmali">${this.localize('interface:vinaya')}</a>
-            </li>
-            <li>
-              <a href="/abhidhamma-guide-sujato">${this.localize('interface:abhidhamma')}</a>
-            </li>
-          `
-        : ''}
-    `;
-  }
-
-  get shouldShowAcademicToolbarTemplate() {
-    return html`
-      ${this.staticPagesToolbarDisplayState &&
-      this.staticPagesToolbarDisplayState.displayAcademicToolbar
-        ? html`
-            <li>
-              <a href="/numbering">${this.localize('interface:numbering')}</a>
-            </li>
-            <li>
-              <a href="/abbreviations">${this.localize('interface:abbreviations')}</a>
-            </li>
-            <li>
-              <a href="/methodology">${this.localize('interface:methodology')}</a>
-            </li>
-          `
-        : ''}
-    `;
-  }
-
-  get shouldShowOrganizationalToolbarTemplate() {
-    return html`
-      ${this.staticPagesToolbarDisplayState &&
-      this.staticPagesToolbarDisplayState.displayOrganizationalToolbar
-        ? html`
-            <li>
-              <a href="/acknowledgments">${this.localize('interface:acknowledgments')}</a>
-            </li>
-            <li>
-              <a href="/licensing">${this.localize('interface:licensing')}</a>
-            </li>
-            <li>
-              <a href="/about">${this.localize('interface:about')}</a>
-            </li>
-          `
-        : ''}
-    `;
-  }
-
-  get shouldShowGuidesToolbarTemplate() {
-    return html`
-      ${this.staticPagesToolbarDisplayState &&
-      this.staticPagesToolbarDisplayState.displayGuidesToolbar
-        ? html`
-            <li>
-              <a href="/general-guide-sujato">${this.localize('interface:general')}</a>
-            </li>
-            <li>
-              <a href="/dn-guide-sujato">${this.localize('interface:long')}</a>
-            </li>
-            <li>
-              <a href="/mn-guide-sujato">${this.localize('interface:middle')}</a>
-            </li>
-            <li>
-              <a href="/sn-guide-sujato">${this.localize('interface:linked')}</a>
-            </li>
-            <li>
-              <a href="/an-guide-sujato">${this.localize('interface:numbered')}</a>
-            </li>
-          `
-        : ''}
-    `;
-  }
-
-  _removeSelectedClass() {
-    this.querySelectorAll('.staticPageSelected').forEach(e => {
-      e.classList.remove('staticPageSelected');
-    });
-  }
-
-  _addSelectedClass(e) {
-    e.classList.add('staticPageSelected');
-  }
-
-  _addStaticPageLinkEventListener() {
-    this.querySelectorAll('#static_pages_nav_menu nav li a').forEach(element => {
-      element.addEventListener('click', e => {
-        this._removeSelectedClass();
-        this._addSelectedClass(element);
-      });
-    });
   }
 
   static get properties() {
@@ -230,7 +66,6 @@ class SCSiteLayout extends LitLocalized(LitElement) {
       changedRoute: { type: Object },
       displaySettingMenu: { type: Boolean },
       toolbarTitle: { type: String },
-      staticPagesToolbarDisplayState: { type: Object },
       linearProgressActive: { type: Boolean },
       toolbarPosition: { type: Object },
     };
@@ -251,17 +86,6 @@ class SCSiteLayout extends LitLocalized(LitElement) {
     this.displaySettingMenu = state.displaySettingMenu;
     this.toolbarTitle = state.toolbarOptions.title;
     this.shouldShowFirstToolbar = true;
-    this.staticPagesToolbarDisplayState = state.staticPagesToolbarDisplayState;
-    if (!this.staticPagesToolbarDisplayState) {
-      this.staticPagesToolbarDisplayState = {
-        displayFirstToolbar: true,
-        displaySecondToolbar: false,
-        displayTipitakaToolbar: false,
-        displayAcademicToolbar: false,
-        displayOrganizationalToolbar: false,
-        displayGuidesToolbar: false,
-      };
-    }
     this.linearProgressActive = false;
     this.toolbarPosition = state.toolbarPosition;
   }
@@ -284,12 +108,6 @@ class SCSiteLayout extends LitLocalized(LitElement) {
         store.dispatch({
           type: 'SET_NAVIGATION',
           navigationArray: navArray,
-        });
-      },
-      setStaticPagesToolbarDisplayState(toolbarDisplayState) {
-        store.dispatch({
-          type: 'CHANGE_STATIC_PAGES_TOOLBAR_DISPLAY_STATE',
-          staticPagesToolbarDisplayState: toolbarDisplayState,
         });
       },
       changeDisplaySuttaParallelsState(displayState) {
@@ -330,9 +148,6 @@ class SCSiteLayout extends LitLocalized(LitElement) {
     if (this.appColorTheme !== state.colorTheme) {
       this.appColorTheme = state.colorTheme;
     }
-    if (this.staticPagesToolbarDisplayState !== state.staticPagesToolbarDisplayState) {
-      this.staticPagesToolbarDisplayState = state.staticPagesToolbarDisplayState;
-    }
     if (this.changedRoute !== state.currentRoute) {
       this.changedRoute = state.currentRoute;
     }
@@ -356,7 +171,7 @@ class SCSiteLayout extends LitLocalized(LitElement) {
     const setScrollbarWidth = () =>
       document.body.style.setProperty(
         '--scrollbar-width',
-        window.innerWidth - document.documentElement.clientWidth + 'px'
+        `${window.innerWidth - document.documentElement.clientWidth}px`
       );
 
     window.addEventListener('resize', setScrollbarWidth, { passive: true });
@@ -378,21 +193,21 @@ class SCSiteLayout extends LitLocalized(LitElement) {
       });
     });
 
-    this.addEventListener('hide-sc-top-sheet', e => {
+    this.addEventListener('hide-sc-top-sheet', () => {
       this.querySelector('#setting_menu')?.hide();
       this.actions.changeDisplaySettingMenuState(false);
     });
 
-    this.addEventListener('show-sc-top-sheet', e => {
+    this.addEventListener('show-sc-top-sheet', () => {
       this.querySelector('#setting_menu')?.show();
     });
 
-    this.addEventListener('hide-sc-sutta-parallels', e => {
+    this.addEventListener('hide-sc-sutta-parallels', () => {
       this.querySelector('#sutta_parallels')?.hide();
       this.actions.changeDisplaySuttaParallelsState(false);
     });
 
-    this.addEventListener('show-sc-sutta-parallels', e => {
+    this.addEventListener('show-sc-sutta-parallels', () => {
       this.querySelector('#sutta_parallels')?.show();
     });
 
@@ -400,12 +215,12 @@ class SCSiteLayout extends LitLocalized(LitElement) {
       this.querySelector('#sutta_parallels').suttaplexItem = e.detail.suttaplexItem;
     });
 
-    this.addEventListener('hide-sc-sutta-toc', e => {
+    this.addEventListener('hide-sc-sutta-toc', () => {
       this.querySelector('#sutta_toc')?.hide();
       this.actions.changeDisplaySuttaToCState(false);
     });
 
-    this.addEventListener('show-sc-sutta-toc', e => {
+    this.addEventListener('show-sc-sutta-toc', () => {
       this.querySelector('#sutta_toc')?.show();
     });
 
@@ -417,7 +232,7 @@ class SCSiteLayout extends LitLocalized(LitElement) {
       }
     });
 
-    this.addEventListener('hide-sc-sutta-info', e => {
+    this.addEventListener('hide-sc-sutta-info', () => {
       this.querySelector('#sutta-info')?.hide();
       const pubInfo = this.querySelector('#bilara-sutta-info');
       if (pubInfo && pubInfo.hide) {
@@ -491,9 +306,6 @@ class SCSiteLayout extends LitLocalized(LitElement) {
     });
 
     this._initNavigation();
-    this._initStaticPagesToolbarDisplayState();
-    this._addStaticPageLinkEventListener();
-    this._setStaticPageMenuItemSelected();
     this._setToolbarPosition();
 
     window.addEventListener('error', e => {
@@ -528,14 +340,6 @@ class SCSiteLayout extends LitLocalized(LitElement) {
     this.getElementById('subTitle').style.transform = 'scale(1)';
   }
 
-  _setStaticPageMenuItemSelected() {
-    this._removeSelectedClass();
-    const element = this.querySelector(`nav a[href="${this.changedRoute.path}"]`);
-    if (element) {
-      element.classList.add('staticPageSelected');
-    }
-  }
-
   _initNavigation() {
     this.navArray = store.getState().navigationArray;
     if (!this.navArray) {
@@ -552,17 +356,6 @@ class SCSiteLayout extends LitLocalized(LitElement) {
     }
   }
 
-  _initStaticPagesToolbarDisplayState() {
-    this.actions.setStaticPagesToolbarDisplayState({
-      displayFirstToolbar: true,
-      displaySecondToolbar: false,
-      displayTipitakaToolbar: false,
-      displayAcademicToolbar: false,
-      displayOrganizationalToolbar: false,
-      displayGuidesToolbar: false,
-    });
-  }
-
   updated(changedProps) {
     if (changedProps.has('siteLanguage')) {
       this._setSiteLanguage();
@@ -572,11 +365,6 @@ class SCSiteLayout extends LitLocalized(LitElement) {
     }
     if (changedProps.has('changedRoute')) {
       this._routeChanged();
-      this._setStaticPageMenuItemSelected();
-    }
-    if (changedProps.has('staticPagesToolbarDisplayState')) {
-      this._addStaticPageLinkEventListener();
-      this._setStaticPageMenuItemSelected();
     }
     if (changedProps.has('toolbarPosition')) {
       this._setToolbarPosition();
