+++ client/elements/static/sc-static-introduction.js
@@ -7,6 +7,10 @@ import { typographyStaticStyles } from '../styles/sc-typography-static-styles';
 import { SCStaticPage } from '../addons/sc-static-page';
 
 class SCStaticIntroduction extends SCStaticPage {
+  createRenderRoot() {
+    return this;
+  }
+
   render() {
     return html`
       <style>
